Fix typing issues of StringSetFlag
* Selectively ignore override errors as this is a custom class
* Use a separate interface declaration file (.pyi) by refactoring out
  StringSetFlag class from the utils module to overcome limitation of
  mypy semantic analysis on "__ror__ = __or__" in subclasses of
  Enum/Flag classes.
